pyproject.toml
[tool.poetry]
name = "testovoe"
version = "0.1.0"
description = ""
authors = ["Sergey <one6sergey@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
ruff = "^0.9.1"
punq = "^0.7.0"
sqlalchemy = "^2.0.37"
pydantic-settings = "^2.7.1"
uvicorn = "^0.34.0"
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
